[CODEGEN][LLVM] Refactor cpu runtime related code to CodeGenCPU
