+++ src/codegen/llvm/codegen_llvm.h
@@ -44,26 +44,26 @@ class CodeGenLLVM :
    * \param dynamic_lookup Whether dynamically lookup runtime function
    *                       or use the runtime function table passed by caller.
    */
-  void Init(const std::string& module_name,
-            llvm::TargetMachine* tm,
-            llvm::LLVMContext* ctx,
-            bool system_lib,
-            bool dynamic_lookup);
+  virtual void Init(const std::string& module_name,
+                    llvm::TargetMachine* tm,
+                    llvm::LLVMContext* ctx,
+                    bool system_lib,
+                    bool dynamic_lookup);
   /*!
    * \brief Compile and add function f to the current module.
    * \param f The function to be added.
    */
-  void AddFunction(const LoweredFunc& f);
+  virtual void AddFunction(const LoweredFunc& f);
   /*!
    * \brief Add main function as the entry name
    * \param entry_func_name The name of entry function to be added.
    */
-  void AddMainFunction(const std::string& entry_func_name);
+  virtual void AddMainFunction(const std::string& entry_func_name);
   /*!
    * \brief Finish current pass of codegen, get the module.
    * \return the created module.
    */
-  std::unique_ptr<llvm::Module> Finish();
+  virtual std::unique_ptr<llvm::Module> Finish();
   /*!
    * \brief Create Value for expression e
    * \param e The expression to be created value for.
@@ -120,8 +120,6 @@ class CodeGenLLVM :
   virtual llvm::Value* CreateIntrinsic(const Call* op);
   // create extern function call
   virtual llvm::Value* CreateCallExtern(const Call* op);
-  // create call into tvm packed function.
-  virtual llvm::Value* CreateCallPacked(const Call* op);
   // Scalarize e by iterating elements of e.
   // f is a callback that takes index and v.
   virtual void Scalarize(const Expr& e,
@@ -134,6 +132,14 @@ class CodeGenLLVM :
     /*! \brief The alignment of allocation */
     int alignment{0};
   };
+  // Initialize target
+  virtual void InitTarget(llvm::TargetMachine* tm);
+  // Add module startup function if needed.
+  virtual void AddStartupFunction() {}
+  // apply optimization on the module.
+  virtual void Optimize();
+  // Get the maximim storage align bits of buffer pointer given storage scope.
+  virtual int NativeVectorBits(const std::string& storage_scope) const;
   /*!
    * \param t The original type.
    * \return LLVM type of t
@@ -145,15 +151,36 @@ class CodeGenLLVM :
   void GetAlignment(
       Type t, const Variable* buf_var, const Expr& index,
       int* p_alignment, int* p_native_bits);
+  // Get constant string
+  llvm::Value* GetConstString(const std::string& str);
   // do a scalarize call with f
   llvm::Value* CreateScalarizedCall(
       const Call* op, llvm::Function* f, const std::vector<llvm::Value*>& args);
-  // Initialize target
-  virtual void InitTarget(llvm::TargetMachine* tm);
-  // apply optimization on the module.
-  virtual void Optimize();
-  // Get the maximim storage align bits of buffer pointer given storage scope.
-  virtual int NativeVectorBits(const std::string& storage_scope) const;
+  // cast operatpr
+  llvm::Value* CreateCast(Type from, Type to, llvm::Value* value);
+  // comparison op
+  llvm::Value* GetVarValue(const Variable* v) const;
+  llvm::Value* CreateLT(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateLE(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateGT(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateGE(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateAdd(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateSub(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateMul(Type t, llvm::Value* a, llvm::Value* b);
+  llvm::Value* CreateBroadcast(llvm::Value* value, int lanes);
+  llvm::Value* CreateBufferPtr(Type t, llvm::Value* buffer, llvm::Value* index);
+  // Vector concatenation.
+  llvm::Value* CreateVecSlice(llvm::Value* vec, int begin, int extent);
+  llvm::Value* CreateVecFlip(llvm::Value* vec);
+  llvm::Value* CreateVecConcat(std::vector<llvm::Value*> vecs);
+  llvm::Value* CreateVecPad(llvm::Value* vec, int target_lanes);
+  // Create serial for
+  void CreateSerialFor(llvm::Value* begin,
+                       llvm::Value* end,
+                       llvm::Value* stride,
+                       const VarExpr& loop_var, const Stmt& body);
+  // add alias information.
+  void AddAliasInfo(llvm::Instruction* load, const Variable* buffer, Expr index, Type type);
   // The IRBuilder.
   using IRBuilder = llvm::IRBuilder<llvm::ConstantFolder, llvm::IRBuilderDefaultInserter>;
   // The current function
@@ -177,129 +204,25 @@ class CodeGenLLVM :
   llvm::Type* t_int32_{nullptr};
   llvm::Type* t_int64_{nullptr};
   llvm::Type* t_float64_{nullptr};
-  // branch
+  // meta data
   llvm::MDNode* md_very_likely_branch_{nullptr};
   llvm::MDNode* md_tbaa_root_{nullptr};
   llvm::MDNode* md_tbaa_alias_set_{nullptr};
-  llvm::MDNode* md_tbaa_ctx_ptr_{nullptr};
-  // TVM related data types
-  llvm::Type* t_tvm_shape_index_{nullptr};
-  llvm::Type* t_tvm_func_handle_{nullptr};
-  llvm::StructType* t_tvm_context_{nullptr};
-  llvm::StructType* t_tvm_type_{nullptr};
-  llvm::StructType* t_tvm_array_{nullptr};
-  llvm::StructType* t_tvm_value_{nullptr};
-  llvm::StructType* t_tvm_parallel_group_env_{nullptr};
-  llvm::FunctionType* ftype_tvm_parallel_lambda_{nullptr};
-  llvm::FunctionType* ftype_tvm_func_call_{nullptr};
-  llvm::FunctionType* ftype_tvm_get_func_from_env_{nullptr};
-  llvm::FunctionType* ftype_tvm_api_set_last_error_{nullptr};
-  llvm::FunctionType* ftype_tvm_parallel_launch_{nullptr};
-  llvm::FunctionType* ftype_tvm_parallel_barrier_{nullptr};
-  llvm::FunctionType* ftype_tvm_register_system_symbol_{nullptr};
-  // Lazy entry for function call.
-  llvm::FunctionType* ftype_tvm_static_init_callback_{nullptr};
-  llvm::FunctionType* ftype_tvm_static_init_{nullptr};
-  // The acting body
-  llvm::BasicBlock* block_{nullptr};
+
   /*! \brief native vector bits of current targetx*/
   int native_vector_bits_{0};
   /*! \brief the storage scope of allocation */
   std::unordered_map<const Variable*, StorageInfo> alloc_storage_info_;
-
- private:
-  // the parallel group information
-  struct ParallelEnv {
-    VarExpr task_id;
-    VarExpr num_task;
-    bool stride_pattern{false};
-    bool hit_parallel_loop{false};
-    llvm::Value* penv{nullptr};
-  };
-  // Get runtime functions
-  llvm::GlobalVariable* InitContextPtr(llvm::Type* type, std::string name);
-  llvm::Value* GetContextPtr(llvm::GlobalVariable* gv);
-  llvm::Value* RuntimeTVMFuncCall();
-  llvm::Value* RuntimeTVMGetFuncFromEnv();
-  llvm::Value* RuntimeTVMAPISetLastError();
-  llvm::Value* RuntimeTVMParallelLaunch();
-  llvm::Value* RuntimeTVMParallelBarrier();
-  // comparison op
-  llvm::Value* GetVarValue(const Variable* v) const;
-  llvm::Value* CreateLT(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateLE(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateGT(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateGE(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateAdd(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateSub(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateMul(Type t, llvm::Value* a, llvm::Value* b);
-  llvm::Value* CreateBroadcast(llvm::Value* value, int lanes);
-  llvm::Value* GetConstString(const std::string& str);
-  llvm::Value* CreateBufferPtr(Type t, llvm::Value* buffer, llvm::Value* index);
-  llvm::Value* CreateStructRefPtr(Type t, llvm::Value* buffer, llvm::Value* index, int kind);
-  llvm::Value* CreateCast(Type from, Type to, llvm::Value* value);
-  llvm::Value* GetPackedFuncHandle(const std::string& str);
-  // Vector concatenation.
-  llvm::Value* CreateVecSlice(llvm::Value* vec, int begin, int extent);
-  llvm::Value* CreateVecFlip(llvm::Value* vec);
-  llvm::Value* CreateVecConcat(std::vector<llvm::Value*> vecs);
-  llvm::Value* CreateVecPad(llvm::Value* vec, int target_lanes);
-  llvm::Value* PackClosureData(const Array<Var>& fields);
-  void UnpackClosureData(llvm::Value*cdata,
-                         const Array<Var>& fields,
-                         std::unordered_map<const Variable*, llvm::Value*>* vmap);
-  // Create static initialization
-  void CreateStaticInit(const std::string& init_fname, const Stmt& body);
-  // Create parallel launch
-  void CreateParallelLaunch(const Stmt& body, int num_task);
-  // Create serial for
-  void CreateSerialFor(llvm::Value* begin,
-                       llvm::Value* end,
-                       llvm::Value* stride,
-                       const VarExpr& loop_var, const Stmt& body);
-  // Create a new compute scope.
-  void CreateComputeScope(const AttrStmt* op);
-  // Check if the call to packed function is successful
-  // if not directly finalize function and pass on return code.
-  // return the end block after the check
-  llvm::BasicBlock* CheckCallSuccess(llvm::Value* retcode);
-  // Add a function to set global module context
-  void InitGlobalContext(bool dynamic_lookup);
-  // Add module startup function if needed.
-  void AddStartupFunction();
-  // add alias information.
-  void AddAliasInfo(llvm::Instruction* load, const Variable* buffer, Expr index, Type type);
   // The definition of local variable.
   std::unordered_map<const Variable*, llvm::Value*> var_map_;
   // global strings
   std::unordered_map<std::string, llvm::Constant*> str_map_;
-  // The alignment information
-  std::unordered_map<const Variable*, arith::ModularEntry> align_map_;
   // Whether current function is restricted
   bool is_restricted_{true};
+  // The alignment information
+  std::unordered_map<const Variable*, arith::ModularEntry> align_map_;
   // set of var that are not restricted(can alias)
   std::unordered_set<const Variable*> alias_var_set_;
-  // Context for injection lookup
-  llvm::GlobalVariable* gv_mod_ctx_{nullptr};
-  llvm::GlobalVariable* gv_tvm_func_call_{nullptr};
-  llvm::GlobalVariable* gv_tvm_get_func_from_env_{nullptr};
-  llvm::GlobalVariable* gv_tvm_api_set_last_error_{nullptr};
-  llvm::GlobalVariable* gv_tvm_parallel_launch_{nullptr};
-  llvm::GlobalVariable* gv_tvm_parallel_barrier_{nullptr};
-  std::unordered_map<std::string, llvm::GlobalVariable*> gv_func_map_;
-  // context for direct dynamic lookup
-  llvm::Function* f_tvm_func_call_{nullptr};
-  llvm::Function* f_tvm_get_func_from_env_{nullptr};
-  llvm::Function* f_tvm_api_set_last_error_{nullptr};
-  llvm::Function* f_tvm_parallel_launch_{nullptr};
-  llvm::Function* f_tvm_parallel_barrier_{nullptr};
-  llvm::Function* f_tvm_register_system_symbol_{nullptr};
-  // Current parallel environment scope.
-  ParallelEnv parallel_env_;
-  // global to packed function handle
-  std::unordered_map<std::string, llvm::GlobalVariable*> func_handle_map_;
-  // List of symbols to be exported to TVM system lib.
-  std::vector<std::pair<std::string, llvm::Value*> > export_system_symbols_;
 };
 }  // namespace codegen
 }  // namespace tvm
